ver. 0.1.0-a8
FIXED - system tag
FIXED - scrolling page stop for nature of her query (problem with friends and no friends)
FIXED - fixed hidden link on activity
FIXED - fixed shared activity of topic
FIXED - fixed problem to publish emoji
FIXED - fixed delete photos from pages
UPDATE - panel news on ACP
UPDATE - position info of shared activity
UPDATE - style for friends list on profile
UPDATE - style for activity status
UPDATE - style for profile
ADDED - last photo of user on right sidebar on profile
ADDED - birthday's box on right sidebar on activity page
ADDED - post Facebook embed
ADDED - url Website embed
ADDED - removed url from activity if show url website embed
ADDED - chat scroll for old messages

+++ social/social_photo.php
@@ -53,7 +53,7 @@ class social_photo
 	    $this->root_path 				= $root_path;	
 		$this->php_ext 					= $php_ext;
         $this->table_prefix 			= $table_prefix;	
-	    $this->pg_social_path 			= $this->root_path.'ext/pgreca/pgsocial';	
+	    $this->pg_social_path 			= './ext/pgreca/pgsocial/images/';	
 	}
 	
 	/**
@@ -61,13 +61,40 @@ class social_photo
 	*/
 	public function getGallery($wall, $where)
 	{
+		switch($where)
+		{
+			case 'page':
+				$where = 1;
+			break;	
+			default:
+				$where = 0;
+			break;
+		}
+		
 		$sql = "SELECT user_id, username, username_clean, user_colour FROM ".USERS_TABLE." WHERE user_id = '".$wall."'";
 		$result = $this->db->sql_query($sql);
 		$user = $this->db->sql_fetchrow($result);
 		$this->db->sql_freeresult($result);			
 		
 		$limit = 6;
-		$sql = "SELECT *, (SELECT photo_file FROM ".$this->table_prefix."pg_social_photos as cov WHERE user_id = '".$wall."' AND ph.gallery_id = cov.gallery_id AND photo_where = '".$where."' ORDER BY photo_time DESC LIMIT 0, 1) as gallery_cover, COUNT(*) as count FROM ".$this->table_prefix."pg_social_photos as ph WHERE user_id = '".$wall."' AND photo_where = '".$where."' GROUP BY gallery_id";
+		$sql = "SELECT *, 
+				(SELECT photo_file 
+				FROM ".$this->table_prefix."pg_social_photos as cov 
+				WHERE cov.user_id = ph.user_id
+					AND ph.gallery_id = cov.gallery_id 
+					AND ph.photo_where = '".$where."' 
+				ORDER BY photo_time DESC 
+				LIMIT 0, 1) as gallery_cover, 
+				(SELECT COUNT(*)
+				FROM ".$this->table_prefix."pg_social_photos as contt
+				WHERE contt.user_id = ph.user_id
+					AND ph.gallery_id = contt.gallery_id
+					AND ph.photo_where = '".$where."'
+				) as count 
+			FROM ".$this->table_prefix."pg_social_photos as ph 
+			WHERE ph.user_id = '".$wall."' 
+				AND ph.photo_where = '".$where."' 
+			GROUP BY gallery_id";
 		$result = $this->db->sql_query($sql);
 		while($row = $this->db->sql_fetchrow($result))
 		{	
@@ -83,12 +110,14 @@ class social_photo
 					$row['gallery_name'] = $this->user->lang('PG_SOCIAL_COVER');
 				break;				
 			}
+			$count_g = $this->user->lang('PHOTO', 1);
+			if($row['count'] == 0 || $row['count'] > 1) $count_g = $this->user->lang('PHOTO', 2);
 			$this->template->assign_block_vars('social_gallery', array(
 				'GALLERY_ID'		=> $row['gallery_id'],
 				'GALLERY_URL'		=> get_username_string('profile', $user['user_id'], $user['username'], $user['user_colour'])."&gall=".$row['gallery_id'],
 				'GALLERY_NAME'		=> $row['gallery_name'],
-				'GALLERY_COUNT'		=> $row['count'],
-				'PHOTO_COVER'		=> $this->pg_social_path."/images/upload/".$row['gallery_cover'],
+				'GALLERY_COUNT'		=> $row['count']." ".$count_g,
+				'PHOTO_COVER'		=> $this->pg_social_path."upload/".$row['gallery_cover'],
 				'PHOTO_FILE'		=> $row['photo_file'],
 			));
 		}
@@ -125,15 +154,28 @@ class social_photo
 	/**
 	 * Array Photos of gallery
 	*/
-	public function getPhotos($where, $user, $gall)
+	public function getPhotos($where, $last = "gall", $user, $gall = false)
 	{
-		$sql = "SELECT * FROM ".$this->table_prefix."pg_social_photos WHERE gallery_id = '".$gall."' AND user_id = '".$user."' AND photo_where = '".$where."' ORDER BY photo_time DESC";
+		$sqlwhere = '';
+		$sqllimit = '';
+		if($last == "last") 
+		{
+			$sqllimit = " LIMIT 0, 9"; 
+			$block = "last_photos";
+		}
+		else 
+		{		
+			$block = 'social_photos';
+			$sqlwhere = " AND gallery_id = '".$gall."'"; 	
+		}
+		$sql = "SELECT * FROM ".$this->table_prefix."pg_social_photos WHERE user_id = '".$user."' AND photo_where = '".$where."'".$sqlwhere." ORDER BY photo_time DESC".$sqllimit;
 		$result = $this->db->sql_query($sql);
 		while($row = $this->db->sql_fetchrow($result))
 		{
-			$this->template->assign_block_vars('social_photos', array(
+			$this->template->assign_block_vars($block, array(
+				"SQL"			=> $sql,
 				"PHOTO_ID"		=> $row['photo_id'],
-				"PHOTO_FILE"	=> $this->pg_social_path."/images/upload/".$row['photo_file'],
+				"PHOTO_FILE"	=> $this->pg_social_path."upload/".$row['photo_file'],
 			));
 		}		
 	}
@@ -143,6 +185,11 @@ class social_photo
 	*/
 	public function getPhoto($photo, $template = NULL)
 	{
+		$allow_bbcode = $this->config['pg_social_bbcode'];
+		$allow_urls = $this->config['pg_social_url'];
+		$allow_smilies = $this->config['pg_social_smilies'];
+		$flags = (($allow_bbcode) ? OPTION_FLAG_BBCODE : 0) + (($allow_smilies) ? OPTION_FLAG_SMILIES : 0) + (($allow_urls) ? OPTION_FLAG_LINKS : 0);
+		
 		$photoe = explode('#', $photo);
 		$photo = $photoe[0];
 		$sql = "SELECT p.*, (SELECT post_ID FROM ".$this->table_prefix."pg_social_wall_post WHERE post_extra = '".$photo."') as post_id FROM ".$this->table_prefix."pg_social_photos AS p WHERE p.photo_id = '".$photo."'";
@@ -152,22 +199,21 @@ class social_photo
 		
 		if($row['photo_where'] == 0)
 		{
-			$sql = "SELECT user_id, username, username_clean, user_colour, user_avatar, user_avatar_type FROM ".USERS_TABLE." WHERE user_id = '".$row['user_id']."'";
+			$sql = "SELECT user_id, username, username_clean, user_colour, user_avatar, user_avatar_type, '' as page_founder FROM ".USERS_TABLE." WHERE user_id = '".$row['user_id']."'";
 		}
 		else
 		{
-			$sql = "SELECT page_id as user_id, page_username as username, page_username as username_clean, '' as user_colour, page_avatar as user_avatar, '' as user_avatar_type FROM ".$this->table_prefix."pg_social_pages WHERE page_id = '".$row['user_id']."'";
+			$sql = "SELECT page_id as user_id, page_username as username, page_username as username_clean, '' as user_colour, page_avatar as user_avatar, '' as user_avatar_type, page_founder FROM ".$this->table_prefix."pg_social_pages WHERE page_id = '".$row['user_id']."'";
 		}
 		$result = $this->db->sql_query($sql);
 		$user = $this->db->sql_fetchrow($result);
 		
 		if(!$template)
 		{
-			$row['photo_file'] = generate_board_url().$this->pg_social_path."/images/upload/".$row['photo_file'];
+			$row['photo_file'] = generate_board_url()."/ext/pgreca/pgsocial/images/upload/".$row['photo_file'];
 			return $row;
 		}
 		else{
-			$user_avatar = $this->pg_social_helper->social_avatar_thumb($this->user->data['user_avatar'], $this->user->data['user_avatar_type']);
 			$comment = "<span>".$this->pg_social_helper->countAction("comments", $row['post_id'])."</span> ";
 			if($this->pg_social_helper->countAction("comments", $row['post_id']) == 0 || $this->pg_social_helper->countAction("comments", $row['post_id']) > 1)
 			{
@@ -177,37 +223,32 @@ class social_photo
 			{
 				$comment .= $this->user->lang('COMMENT', 1);
 			}
-			$likes = "<span>".$this->pg_social_helper->countAction("like", $row['post_id'])."</span> ";
+			$likes = "<span>".$this->pg_social_helper->countAction("like", $row['post_id'])."</span> ".$this->user->lang('LIKE', 1);
 			if($this->pg_social_helper->countAction("like", $row['post_id']) == 0 | $this->pg_social_helper->countAction("like", $row['post_id']) > 1)
-			{
-				$likes .= $this->user->lang('LIKE', 2);
-			}
-			else
-			{
-				$likes .= $this->user->lang('LIKE', 1);
-			}			
 			
-			if($this->user->data['user_id'] == $user['user_id'] && $row['photo_file'] != $this->user->data['user_pg_social_cover']) $photo_action = 1; else $photo_action = 0;	
-				
+			$desc = $this->pg_social_helper->noextra(generate_text_for_display($row['photo_desc'], '', '', $flags));
+			if(($this->user->data['user_id'] == $user['user_id'] && $row['photo_file'] != $this->user->data['user_pg_social_cover']) || $this->user->data['user_id'] == $user['page_founder']) $photo_action = 1; else $photo_action = 0;	
+			if(!array_key_exists('username', $row)) $row['username'] = "";
+			if(!array_key_exists('user_colour', $row)) $row['user_colour'] = "";
+			($row['username'] && $row['user_colour']) ? $gallery_url = get_username_string('profile', $row['user_id'], $row['username'], $row['user_colour'])."&gall=".$row['gallery_id'] : $gallery_url = '';
 			$this->template->assign_block_vars('social_photo', array(
 				'PHOTO_ID'					=> $row['photo_id'],
-				'PHOTO_FILE'				=> generate_board_url().$this->pg_social_path."/images/upload/".$row['photo_file'],
+				'PHOTO_FILE'				=> generate_board_url().$this->pg_social_path."upload/".$row['photo_file'],
 				'PHOTO_TIME'				=> $this->pg_social_helper->time_ago($row['photo_time']),
 				'PHOTO_ACTION'				=> $photo_action,
 				'AUTHOR_PROFILE'			=> get_username_string('profile', $user['user_id'], $user['username'], $user['user_colour']),
 				'AUTHOR_USERNAME'			=> $user['username'],
 				'AUTHOR_COLOUR'				=> '#'.$user['user_colour'],
-				'AUTHOR_AVATAR'				=> ($row['photo_where'] == 0 ? $this->pg_social_helper->social_avatar_thumb($user['user_avatar'], $user['user_avatar_type']) : '<img src="'.$this->pg_social_path.'/images/'.($page['page_avatar'] != "" ? $page_avatar = 'upload/'.$page['page_avatar'] : $page_avatar = 'page_no_avatar.jpg').'" />'),
-				'GALLERY_URL'				=> get_username_string('profile', $row['user_id'], $row['username'], $row['user_colour'])."&gall=".$row['gallery_id'],
+				'AUTHOR_AVATAR'				=> ($row['photo_where'] == 0 ? $this->pg_social_helper->social_avatar_thumb($user['user_avatar'], $user['user_avatar_type'], $this->user->data['user_avatar_width'], $this->user->data['user_avatar_height']) : '<img src="'.generate_board_url().'/ext/pgreca/pgsocial/images/'.($page['page_avatar'] != "" ? $page_avatar = 'upload/'.$page['page_avatar'] : $page_avatar = 'page_no_avatar.jpg').'" />'),
+				'GALLERY_URL'				=> $gallery_url,
 				'PHOTO_ALBUM'				=> $this->gallery_info($row['gallery_id'])['gallery_name'],
-				'PHOTO_DESC'				=> htmlspecialchars_decode($row['photo_desc']),
+				'PHOTO_DESC'				=> $desc,
 				"LIKE"						=> $likes,
 				"IFLIKE"					=> $this->pg_social_helper->countAction("iflike", $row['post_id']),
 				"PRE"						=> $this->prenextPhoto($row['photo_id'], 0, $row['photo_where'], false),
 				"NEX"						=> $this->prenextPhoto($row['photo_id'], 1, $row['photo_where'], false),
 				"COMMENT"					=> $comment,
 				'POST_ID'					=> $row['post_id'],
-				'USER_AVATAR'				=> $user_avatar,
 			));
 			return $this->helper->render('pg_social_photo.html', '');
 		}
@@ -232,7 +273,7 @@ class social_photo
 	    $ImageSize 		= $photo['size']; 
 	    $TempSrc	 	= $photo['tmp_name']; 
 	    $ImageType	 	= $photo['type']; 
-		$imageAlbum 	= $this->pg_social_path.'/images/upload/';
+		$imageAlbum 	= $this->pg_social_path.'upload/';
 
 		if(!file_exists($imageAlbum)) mkdir($imageAlbum);
 		$BigImageMaxSize 		= 1500; //Image Maximum height or width
@@ -380,8 +421,8 @@ class social_photo
 			return $this->add_statusPhoto($where, $who, $user, $gallery, 1, $row['photo_id'], $msg);
 		}
 	}
-	
 	/**
+	
 	 * New activity for upload photo
 	*/
 	public function add_statusPhoto($where, $who, $user, $type, $privacy, $photo, $text)
@@ -389,18 +430,17 @@ class social_photo
 		$allow_bbcode = $this->config['pg_social_bbcode'];
 		$allow_urls = $this->config['pg_social_url'];
 		$allow_smilies = $this->config['pg_social_smilies'];
-		$text_clear = urldecode($text);
+		$text = urldecode($text);
 		$time = time();
-		if(!$extra) $extra = "";
-		$asds = $this->social_tag->showTag($text_clear);
-		
 		generate_text_for_storage($text, $uid, $bitfield, $flags, $allow_bbcode, $allow_urls, $allow_smilies);
-		
+			
+		$text = str_replace('&amp;nbsp;', ' ', $text);
 		$sql_arr = array(
+			'post_parent'		=> 0,
 			'post_where'		=> $where,
 			'wall_id'			=> $who,
 			'user_id'			=> $user,
-			'message'			=> $asds,
+			'message'			=> $text,
 			'time'				=> $time,
 			'post_privacy'		=> $privacy,
 			'post_type'			=> $type,
@@ -412,9 +452,11 @@ class social_photo
 		$sql = "INSERT INTO " . $this->table_prefix . 'pg_social_wall_post' . $this->db->sql_build_array('INSERT', $sql_arr);
 		if($this->db->sql_query($sql))
 		{	
-			$last_status = "SELECT post_ID FROM ".$this->table_prefix."pg_social_wall_post WHERE time = '".$time."' AND user_id = '".$user_id."' AND wall_id = '".$wall_id."' ORDER BY time DESC LIMIT 0, 1";
+			$sql = "UPDATE ".$this->table_prefix."pg_social_photos SET photo_desc = '".$text."' WHERE photo_id = '".$photo."'";
+			$this->db->sql_query($sql);
+			$last_status = "SELECT post_ID FROM ".$this->table_prefix."pg_social_wall_post WHERE time = '".$time."' AND user_id = '".$this->user->data['user_id']."' AND wall_id = '".$wall_id."' ORDER BY time DESC LIMIT 0, 1";
 			$last = $this->db->sql_query($last_status);
-			$row = $this->db->sql_fetchrow();	
+			$row = $this->db->sql_fetchrow($last);	
 			$this->social_tag->addTag($row['post_ID'], $text_clear);
 		}	
 	}
@@ -423,21 +465,37 @@ class social_photo
 	 * Delete photo
 	*/
 	public function deletePhoto($photo)
-	{
-		$sql = "SELECT photo_id, photo_file, user_id FROM ".$this->table_prefix."pg_social_photos WHERE photo_id = '".$photo."'";
-		$query = $this->db->sql_query($sql);
-		$row = $this->db->sql_fetchrow();
+	{		
+		$delphoto = false;
 		
-		$photo = $row['photo_file'];
-		$file = $this->pg_social_path."/images/upload/".$photo;
-		$deletePhoto = "DELETE FROM ".$this->table_prefix."pg_social_photos WHERE photo_id = '".$row['photo_id']."' AND user_id = '".$this->user->data['user_id']."'";
-		if($this->db->sql_query($deletePhoto) && unlink($file))
-		{ 	
-			$deletePost = "DELETE FROM ".$this->table_prefix."pg_social_wall_post WHERE post_extra = '".$row['photo_id']."'";
-			$this->db->sql_query($deletePost);
-			$this->template->assign_vars(array(
-				"ACTION"	=>  "deleted",
-			));
+		$sql = "SELECT photo_id, photo_file, user_id, photo_where FROM ".$this->table_prefix."pg_social_photos WHERE photo_id = '".$photo."'";
+		$query = $this->db->sql_query($sql);
+		$row = $this->db->sql_fetchrow($query);
+		if($row['photo_where'] == 1) 
+		{
+			$pagesql = "SELECT page_founder FROM ".$this->table_prefix."pg_social_pages WHERE page_id = '".$row['user_id']."'";
+			$pageresult = $this->db->sql_query($pagesql);
+			$rowpage = $this->db->sql_fetchrow($pageresult);
+			if($rowpage['page_founder'] == $this->user->data['user_id']) $delphoto = true;
+		} 
+		elseif($row['user_id'] == $this->user->data['user_id'])
+		{
+			$delphoto = true;
+		}
+		
+		if($delphoto)
+		{
+			$photo = $row['photo_file'];
+			$file = $this->pg_social_path.'upload/'.$photo;
+			$deletePhoto = "DELETE FROM ".$this->table_prefix."pg_social_photos WHERE photo_id = '".$row['photo_id']."'";
+			if($this->db->sql_query($deletePhoto) && unlink($file))
+			{ 	
+				$deletePost = "DELETE FROM ".$this->table_prefix."pg_social_wall_post WHERE post_extra = '".$row['photo_id']."'";
+				$this->db->sql_query($deletePost);
+				$this->template->assign_vars(array(
+					"ACTION"	=>  "deleted",
+				));
+			}
 		}
 		return $this->helper->render('activity_status_action.html', "");
 	}
@@ -457,9 +515,8 @@ class social_photo
 			$orde = "<";
 			$ordn = "DESC";
 		}
-		if($where == "profile") $wher = '0'; elseif($where == "page") $wher = '1';
 		$photo_info = $this->getPhoto($photo);
-		$sql = "SELECT photo_id FROM ".$this->table_prefix."pg_social_photos WHERE photo_id ".$orde." '".$photo_info['photo_id']."' AND photo_where = '".$where."' AND user_id = '".$photo_info['user_id']."' AND gallery_id = '".$photo_info['gallery_id']."' ORDER BY photo_id ".$ordn." LIMIT 0, 1"; 
+		$sql = "SELECT photo_id FROM ".$this->table_prefix."pg_social_photos WHERE photo_id ".$orde." '".$photo_info['photo_id']."' AND photo_where = '".$photo_info['photo_where']."' AND user_id = '".$photo_info['user_id']."' AND gallery_id = '".$photo_info['gallery_id']."' ORDER BY photo_id ".$ordn." LIMIT 0, 1"; 
 		$result = $this->db->sql_query($sql);
 		$row = $this->db->sql_fetchrow($result);
